requirements.txt
pyqt6==6.7.1
pymongo==4.11.1
python-dotenv==1.0.1
#pyinstaller
